feat: Add input mode switcher for physical keyboard
Adds an input mode switcher to toggle between Cantonese and English (ABC) modes when using a physical keyboard.

The switcher is accessible via a new button on the candidate bar and by pressing the Shift key.

The physical keyboard's collapsed candidate view is refactored into a new `PhysicalKeyboardCandidateBar` composable, which fixes issue that the UI of the bar is not align with the virtrual keyboard's bar